pyproject.toml
[project]
name = "cheltenham"
version = "0.1.0"
description = "Cheltenham Festival 2026 value finder"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.0",
    "pyarrow>=18.0.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "lightgbm>=4.5.0",
    "scikit-learn>=1.6.0",
    "matplotlib>=3.9.0",
    "numpy>=2.1.0",
    "anthropic>=0.42.0",
    "requests>=2.32.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"
